I2C scan should write, not read. (Needs QA).
Teensy code was scanning using write; not all devices respond to reads,
(eg: Disting NX) and as such, this was failing.

Fortunately, the Pico SDK documentation suggests we can use response to
writes - either a failure or result - as a way of detecting if devices
exist on the bus.

This has not been tested yet, and needs evaluating with devices.